+++ src/test/java/be/ugent/rml/Mapper_HTML_Test.java
@@ -7,34 +7,26 @@ import org.junit.runners.Parameterized;
 import java.util.Arrays;
 
 @RunWith(Parameterized.class)
-public class Mapper_CSVW_Test extends TestCore {
+public class Mapper_HTML_Test extends TestCore {
     @Parameterized.Parameter(0)
     public String testCaseName;
 
     @Parameterized.Parameters(name = "{index}: CSVW_{0}")
     public static Iterable<Object[]> data() {
         return Arrays.asList(new Object[][]{
-                {"delimiter"},
-                {"encoding"},
-                {"tabs"},
-                {"tabs_unicode"},
-                {"trim"}
+                {"RMLTC0001a"},
+                {"RMLTC0015a"}
         });
 
     }
 
     @Test
     public void doMapping() throws Exception {
-
-//        //setup expected exception
-//        if (expectedException != null) {
-//            thrown.expect(expectedException);
-//        }
         doCSVWTest(testCaseName);
     }
 
     private void doCSVWTest(String name) {
-        String base = "./test-cases-CSVW/RMLTC0002a_" + name + "-CSVW/";
+        String base = "./test-cases-HTML/" + name + "-HTML/";
         doMapping(base + "mapping.ttl", base + "output.nq");
     }
 }